Trying new version of minimal_polynomial that looks for lower degree polynomials first.
I'm still not sure that this is guranteed to be correct. There appears to be a lot of confusion in the literature as to exactly how many decimal places you need in order to get the minimal polynomial. However, this change appears to fix the cases found by Shannon Horrigan.