Added __slots__ to Vertex, Edge, Triangle and Corner.
This appears to have had a pretty dramatic effect on the memory usage.
Reducing it by more than a factor of two and so the "hard" examples
found by Nathan now stay in memory and so run ~10x faster.

These will need to be updated if additional properties are added to
any of these classes in the future.